test: move the import.meta.main compile cases to their own file
bundler_compile.test.ts takes 223 s to 227 s on Windows 11 arm64 on main.
The test runner stops a bundler_compile file at 300 s. With these nine
cases, one CI run took 246 s, and the next run went past the limit on all
four attempts. bundler_compile_import_meta_main.test.ts gets its own limit,
and bundler_compile.test.ts is the same as on main again.
